requirements.txt
asgiref==3.4.1
autopep8==1.5.7
dj-database-url==0.5.0
Django==3.2.6
django-bootstrap-static==4.2.1
django-crispy-forms==1.13.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
Pillow==8.4.0
pycodestyle==2.7.0
PyJWT==2.3.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.2
Unipath==1.1
whitenoise==5.3.0
